+++ transfunctions/__init__.py
@@ -1,4 +1,4 @@
-from transfunctions.decorator import transfunction as transfunction
+from transfunctions.decorators.transfunction import transfunction as transfunction
 from transfunctions.decorators.superfunction import superfunction as superfunction
 
 from transfunctions.markers import (
